third_party/blink/web_tests/http/tests/devtools/profiler/heap-snapshot-summary-search-by-id-expected.txt

Tests search in Summary view of detailed heap snapshots.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testSearch
PASS: hidden node @1 was not found.
PASS: found node @101 with class 'A'
PASS: node @a was not found
PASS: node @999 was not found
PASS: found node @100 with class 'B'
PASS: found node @400 with class 'B'
PASS: found node @401 with class 'A'

Profiler was disabled.
